Fixed to be accurate and follow the binary XY sex-determination system
The XY sex-determination system is the sex-determination system found in humans, most other mammals, some insects (Drosophila), some snakes, and some plants (Ginkgo). In this system, the sex of an individual is determined by a pair of sex chromosomes (gonosomes). Females typically have two of the same kind of sex chromosome (XX), and are called the homogametic sex. Males typically have two distinct sex chromosomes (XY), and are called the heterogametic sex. Exceptions to these generalisations happen in the cases of XX males or XY females, or other syndromes.

The XY system contrasts in several ways with the ZW sex-determination system found in birds, some insects, many reptiles, and various other animals, in which the heterogametic sex is female. It had been thought for several decades that in all snakes gender was determined by the ZW system, but there had been observations of unexpected effects in the genetics of species in the families Boidae and Pythonidae; for example, parthenogenic reproduction produced only females rather than males, which is the opposite of what is to be expected in the ZW system. In the early years of the 21st century such observations prompted research that demonstrated that all pythons and boas so far investigated definitely have the XY system of sex determination.

All animals have a set of DNA coding for genes present on chromosomes. In humans, most mammals, and some other species, two of the chromosomes, called the X chromosome and Y chromosome, code for sex. In these species, one or more genes are present on their Y-chromosome that determine maleness. In this process, an X chromosome and a Y chromosome act to determine the sex of offspring, often due to genes located on the Y chromosome that code for maleness. Offspring have two sex chromosomes: an offspring with two X chromosomes will develop female characteristics, and an offspring with an X and a Y chromosome will develop male characteristics.

## Humans

Human male XY chromosomes after G-banding

In humans, half of spermatozoons carry X chromosome and the other half Y chromosome. A single gene (SRY) present on the Y chromosome acts as a signal to set the developmental pathway towards maleness. Presence of this gene starts off the process of virilization. This and other factors result in the sex differences in humans. The cells in females, with two X chromosomes, undergo X-inactivation, in which one of the two X chromosomes is inactivated. The inactivated X chromosome remains within a cell as a Barr body.

Humans, as well as some other organisms, can have a chromosomal arrangement that is contrary to their phenotypic sex; for example, XX males or XY females (see androgen insensitivity syndrome). Additionally, an abnormal number of sex chromosomes (aneuploidy) may be present, such as Turner's syndrome, in which a single X chromosome is present, and Klinefelter's syndrome, in which two X chromosomes and a Y chromosome are present, XYY syndrome and XXYY syndrome. Other less common chromosomal arrangements include: triple X syndrome, 48, XXXX, and 49, XXXXX.

http://www.bio.uu.nl/endocrinology/teaching/Sex%20Determinatie.pdf